Handle external routes partially
