+++ core-ui/src/components/App/App.js
@@ -18,6 +18,7 @@ import { Sidebar } from 'sidebar/Sidebar';
 import { useInitTheme } from './useInitTheme';
 import { useAuthHandler } from 'state/authDataAtom';
 import { useGetConfiguration } from 'state/configurationAtom';
+import { useGetExtensions } from 'state/navigation/extensibilityNodeAtom';
 
 import ClusterList from 'components/Clusters/views/ClusterList';
 import ClusterRoutes from './ClusterRoutes';
@@ -35,6 +36,7 @@ export default function App() {
   useInitTheme();
   useAuthHandler();
   useGetConfiguration();
+  useGetExtensions();
 
   useEffect(() => {
     i18n.changeLanguage(language);
