+++ core-ui/src/components/App/ClusterRoutes.js
@@ -4,7 +4,6 @@ import { useTranslation } from 'react-i18next';
 import { useRecoilValue, useRecoilState } from 'recoil';
 import { Navigate, useNavigate, useParams } from 'react-router-dom';
 
-import { useMicrofrontendContext } from 'shared/contexts/MicrofrontendContext';
 import { WithTitle } from 'shared/hooks/useWindowTitle';
 import { ClusterOverview } from 'components/Clusters/views/ClusterOverview/ClusterOverview';
 import { clusterState } from 'state/clusterAtom';
@@ -15,15 +14,16 @@ import NamespaceRoutes from './NamespaceRoutes';
 import { otherRoutes } from 'resources/other';
 import { resourceRoutes } from 'resources';
 import { createExtensibilityRoutes } from './ExtensibilityRoutes';
+import { extensibilityNodesState } from 'state/navigation/extensibilityNodeAtom';
 
 export default function ClusterRoutes() {
   let { currentClusterName } = useParams() || {};
 
   const navigate = useNavigate();
-  const { customResources = [] } = useMicrofrontendContext();
   const { t } = useTranslation();
   const language = useRecoilValue(languageAtom);
   const clusters = useRecoilValue(clustersState);
+  const extensions = useRecoilValue(extensibilityNodesState);
   const [cluster, setCluster] = useRecoilState(clusterState);
 
   useEffect(() => {
@@ -55,7 +55,7 @@ export default function ClusterRoutes() {
       />
       <Route path="namespaces/:namespaceId/*" element={<NamespaceRoutes />} />
       {/* extensibility routes should go first, so if someone overwrites the default view, the new one should have a higher priority */}
-      {customResources?.map(cr => createExtensibilityRoutes(cr, language))}
+      {extensions?.map(cr => createExtensibilityRoutes(cr, language))}
       {resourceRoutes}
       {otherRoutes}
     </Routes>
