+++ core-ui/src/components/App/ExtensibilityRoutes.js
@@ -9,8 +9,8 @@ const Details = React.lazy(() =>
   import('../Extensibility/ExtensibilityDetails'),
 );
 
-export const createExtensibilityRoutes = (cr, language, namespaced) => {
-  const urlPath = cr?.general?.urlPath;
+export const createExtensibilityRoutes = (cr, language) => {
+  const urlPath = cr?.pathSegment;
 
   const translationBundle = urlPath || 'extensibility';
   i18next.addResourceBundle(
@@ -19,55 +19,28 @@ export const createExtensibilityRoutes = (cr, language, namespaced) => {
     cr?.translations?.[language] || {},
   );
 
-  if (cr.general?.scope === 'namespace' && namespaced) {
-    return (
-      <React.Fragment key={`namespace-${urlPath}`}>
+  return (
+    <React.Fragment key={urlPath}>
+      <Route
+        path={urlPath}
+        exact
+        element={
+          <Suspense fallback={<Spinner />}>
+            <List />
+          </Suspense>
+        }
+      />
+      {cr.details && (
         <Route
-          path={urlPath}
+          path={`${urlPath}/:resourceName`}
           exact
           element={
             <Suspense fallback={<Spinner />}>
-              <List />
+              <Details />
             </Suspense>
           }
         />
-        {cr.details && (
-          <Route
-            path={`${urlPath}/:resourceName`}
-            exact
-            element={
-              <Suspense fallback={<Spinner />}>
-                <Details />
-              </Suspense>
-            }
-          />
-        )}
-      </React.Fragment>
-    );
-  } else {
-    return (
-      <React.Fragment key={`cluster-${urlPath}`}>
-        <Route
-          path={`${urlPath}`}
-          exact
-          element={
-            <Suspense fallback={<Spinner />}>
-              <List />
-            </Suspense>
-          }
-        />
-        {cr.details && (
-          <Route
-            path={`${urlPath}/:resourceName`}
-            exact
-            element={
-              <Suspense fallback={<Spinner />}>
-                <Details />
-              </Suspense>
-            }
-          />
-        )}
-      </React.Fragment>
-    );
-  }
+      )}
+    </React.Fragment>
+  );
 };
